requirements.txt
# AI-Powered Database Observability System
# Production-ready requirements with version pinning

# Core AI/LLM dependencies
google-generativeai>=0.8.0,<1.0.0

# Environment and configuration
python-dotenv>=1.0.0,<2.0.0

# Web interface
streamlit>=1.32.0,<2.0.0

# Database connectivity
sqlalchemy>=2.0.0,<3.0.0
pyodbc>=5.0.0,<6.0.0

# System monitoring
psutil>=5.9.0,<6.0.0

# XML processing and utilities
lxml>=4.9.0,<5.0.0

# Optional: Additional database drivers
# sqlite3 (built into Python)
# psycopg2-binary>=2.9.0  # PostgreSQL
# pymysql>=1.0.0          # MySQL